requirements.txt
typer

numpy

pytorch